Refactor: dataset validator resource validation errors localtion
